requirements.txt
flask
flask-cors
tensorflow==2.16.1
opencv-python
Pillow
google-generativeai
